Refactor atom-mapping
* Move atom mapping utilities from examples script into fe/atom_mapping.py
* Add get_core_by_smarts
